Huge, mondo checkin.
git-svn-id: http://www.piratestudios.com/pammo/svn@410 fc494e2b-ab4b-0410-a0f4-e0a3d7fb9d21

+++ client/Source/Engine2/enemyWeaponFlamethrower.cpp
@@ -50,7 +50,7 @@ void enemyWeaponFlamethrowerCb(Enemy* e, EnemyWeapon* w, EnemyManager* manager)
             duration = period - duration;
         }
 
-        data->mFireStateChange = now + duration;
+        data->mFireStateChange = now + (uint64_t)duration;
     }
     if(data->mFiring == false)
     {
