+++ client/Source/Engine2/enemyWeaponMachineGun.cpp
@@ -59,7 +59,7 @@ void enemyWeaponMachineGunCb(Enemy* e, EnemyWeapon* w, EnemyManager* manager)
             duration = period - duration;
         }
 
-        data->mFireStateChange = now + duration;
+        data->mFireStateChange = now + (uint64_t)duration;
     }
     if(data->mFiring == false)
     {
